API/get_product_by_description/requirements.txt
Flask
gunicorn
google-cloud-datastore

google-cloud-storage